Split out io.
